Type as int64_t instead

+++ pandas/_libs/tslibs/period.pyx
@@ -1643,7 +1643,7 @@ cdef class _Period(PeriodMixin):
         return freq
 
     @classmethod
-    def _from_ordinal(cls, ordinal: int, freq) -> "Period":
+    def _from_ordinal(cls, ordinal: int64_t, freq) -> "Period":
         """
         Fast creation from an ordinal and freq that are already validated!
         """
@@ -2618,7 +2618,7 @@ class Period(_Period):
                                      dt.hour, dt.minute, dt.second,
                                      dt.microsecond, 1000*nanosecond, base)
 
-        return cls._from_ordinal(int(ordinal), freq)
+        return cls._from_ordinal(ordinal, freq)
 
 
 cdef bint is_period_object(object obj):
